Shared functions enhancement (#843)
* Shared functions enhancement

* run the CI

* rename room session stream and record instances

* rename all the RT methods

* include changeset

* fix start recording test

* remove branch from the CI
