+++ packages/core/src/rooms/index.ts
@@ -9,10 +9,6 @@ export interface BaseRoomInterface<
 }
 
 export * from './methods'
-export * from './methodsRT'
 export * from './RoomSessionRecording'
-export * from './RoomSessionRTRecording'
 export * from './RoomSessionPlayback'
-export * from './RoomSessionRTPlayback'
 export * from './RoomSessionStream'
-export * from './RoomSessionRTStream'
